pyproject.toml
[project]
name = "gemini-pageant"
version = "0.1.0"
description = "Gemini Nano Banana Pro image generation showcase"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]
